+++ src/sccz80/float.h
@@ -4,4 +4,4 @@ extern void fltadd(unsigned char *, unsigned char *);
 extern void fltmult(unsigned char *, unsigned char *);
 extern void pack(unsigned char *, unsigned char *);
 extern void unpack(unsigned char *, unsigned char *);
-extern void norm(unsigned long, unsigned long, unsigned long, int, unsigned char *, int);
+extern void norm(uint32_t, uint32_t, uint32_t, int, unsigned char *, int);
